Add software utility for automatic loading of python libraries
This code allows you to automatically download the necessary libraries 
through the Python programming language's PIP package manager. 
The list of required libraries is given in the file "requirements.txt". 
This code is intended for fast and automatic deployment of already existing 
program code on a new device, automating the process of downloading 
the necessary libraries through the PIP package manager.